requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Neo4j database
neo4j==5.25.0

# Data processing (Python 3.13 compatible versions)
pandas==2.2.3
numpy==2.1.3

# NLP and embeddings (for advanced recommendations)
sentence-transformers==3.2.1
scikit-learn==1.5.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Development
pytest==8.3.3
httpx==0.27.2
